Create brussels3
We are active since september 2018, distributing PM sensors in kit form, utilising the PCB of Influencair.
We are supporting these via e-mail, solving technical problems by e-mail, or in the users home. 
We already distributed more than 3000 sensors in kit form.
We designed a 52 cm long display with 3.3 cm big digits, displaying PM values, temperature en moisture values, together with qualifying messages to expose PM readings to the general public.
The hardware is constructed 4 display blocks of 4 matrix 8x8 displays in cascade, and controlled by an ESP8266.
We intend to put the hardware and software design of the display in public domain, and make ik public on GitHub any time soon, for use by other groups.